update labels type

+++ Residual-Attention-Network/train.py
@@ -130,7 +130,7 @@ def main(args):
                 label_idx_list = sample['labels']
                 batch = torch.stack(sample['preprocessed_images'], dim=0).squeeze()
                 images = batch.to(device)
-                labels = torch.Tensor(label_idx_list).to(device).long()
+                labels = torch.Tensor(label_idx_list).to(device).float()
 
                 # images = Variable(images.cuda())
                 # # print(images.data)
